+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/crediya/api/dto/response/UserResponseDTO.java
@@ -3,6 +3,7 @@ package co.com.crediya.api.dto.response;
 import lombok.Builder;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Builder
 public record UserResponseDTO(
@@ -14,7 +15,7 @@ public record UserResponseDTO(
         String phoneNumber,
         String email,
         String identificationNumber,
-        Double baseSalary,
+        BigDecimal baseSalary,
         RoleDTO role
 ) implements Serializable {
 }
